dataset-generation/pyproject.toml
[project]
name = "neurobench-gen"
version = "0.1.0"
description = "NeuroBench dataset generation pipeline"
requires-python = ">=3.11"
dependencies = [
    "neuroagent-schemas",
    "pydantic>=2.0",
    "pyyaml",
    "rich",
    "typer",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neurobench_gen"]

[tool.uv.sources]
neuroagent-schemas = { workspace = true }
